move asserts to a util file
