+++ src/ast/constants.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 the Skulpt Project
 // SPDX-License-Identifier: MIT
 
-import { assert } from "../parser/pegen.ts";
+import { assert } from "../util/assert.ts";
 
 // deno-lint-ignore no-explicit-any
 export class pyConstant<V = any> {
